+++ src/lib/components/Footer.svelte
@@ -1,5 +1,3 @@
-<footer class="footer footer-center p-4 text-base-content">
-    <aside>
-      <p>5 Luchas Clandestino © 2023 - Todos los derechos reservados</p>
-    </aside>
-  </footer>
+<footer class="flex justify-center p-4 text-base-content">
+	<p>5 Luchas Clandestino © 2023 - Todos los derechos reservados</p>
+</footer>