+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Add
+
+- loading feedback
+
+### Fix
+
+- Remove merch route, only Ordenes
+- migration to shadcn components
+
 ## [0.0.31] - 2026-05-14
 
 ### Add